+++ packages/wise_feedback/lib/src/template/template.dart
@@ -1,0 +1,3 @@
+export 'bug_report_template.dart';
+export 'feedback_field.dart';
+export 'feedback_template.dart';