+++ packages/wise_text/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.2
+
+* Fixed crash where Swift classier would release() twice
+* Changed disposing ownership for `WiseTextClassifier` with custom classifiers
+* `WiseTextWidget` doesn't initialize a classifier anymore when the widget's classified is false
+
 ## 1.0.1
 
 * Bugfixes to FFI calls